requirements.txt
Django
django-snowflake
djangorestframework
requests
snowflake-connector-python
sortedcontainers
sqlparse
django-cors-headers
